Separate MindMapCreate components
